Worked on 0.01$ per 1 second in stoptimer and display the charges in payment fragment

+++ app/src/main/java/ca/group6/it/thedocky/TransactionHistory.java
@@ -2,7 +2,7 @@ package ca.group6.it.thedocky;
 
 public class TransactionHistory {
 
-    public int amount;
+    public float amount;
     public String date;
     public String cityName;
     public String postId;
@@ -10,7 +10,7 @@ public class TransactionHistory {
     public TransactionHistory() {
     }
 
-    public TransactionHistory(int amount, String date, String cityName, String postId) {
+    public TransactionHistory(float amount, String date, String cityName, String postId) {
         this.amount = amount;
         this.date = date;
         this.cityName = cityName;
